// src/types/modal.ts
export interface modalStyleOptions {
  modalBgColor?: string;
  modalBorderRadius?: string;
  buttonBgColor?: string;
  buttonHoverColor?: string;
  buttonTextColor?: string;
  titleTextColor?: string;
  descriptionTextColor?: string;
  buttonCloseColor?: string;
  buttonCloseHoverColor?: string;
}